carousel-studio/backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-multipart==0.0.9
Pillow==10.4.0
httpx==0.27.2
